# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://raapi.test"
        }
    ],
    "paths": {
        "/api/v1/sitios-turisticos/": {
            "get": {
                "tags": [
                    "Sitios turisticos"
                ],
                "summary": "Obtiene todos los sistios turisticos",
                "description": "Regresa los sitios turisticos",
                "operationId": "obtenerSitiosTuristico",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/sitios-turisticos/{id}": {
            "get": {
                "tags": [
                    "Sitios turisticos"
                ],
                "summary": "Mostrar un solo sitio turistico",
                "description": "Regresa el sitio turistico especifico",
                "operationId": "obtenerSitioTuristicoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del sitio turistico",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/v1/sitios-turisticos-visitas/": {
            "post": {
                "tags": [
                    "Sitios turisticos"
                ],
                "summary": "Agrega una visita al sistio turistico",
                "description": "Regresa el sitio turistico",
                "operationId": "agregarSitioTuristicoVisita",
                "parameters": [
                    {
                        "name": "dato",
                        "in": "path",
                        "description": "Objeto JSON con datos de la visita",
                        "required": true,
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "default": "sitios-turisticos-visitas",
                                            "example": "sitios-turisticos-visitas"
                                        },
                                        "attributes": {
                                            "properties": {
                                                "sitio_turistico_id": {
                                                    "type": "number",
                                                    "default": 1,
                                                    "example": 1
                                                },
                                                "ip": {
                                                    "type": "string",
                                                    "example": "192.168.1.1"
                                                },
                                                "fecha": {
                                                    "type": "datetime",
                                                    "example": "2021-11-19 13:39:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the customer"
                    },
                    "400": {
                        "description": "Email required"
                    },
                    "404": {
                        "description": "Customer does not exist"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}